[fix] 프로필 기본URL일때 이미지삭제함수 미실행 수정

+++ middlewares/profileMulter.js
@@ -61,6 +61,8 @@ const uploadProfile = multer({
 });
 
 const deleteProfile = (url) => {
+    if (url === 'https://www.snsboom.co.kr/common/img/default_profile.png')
+        return;
     const filename = url.split('/')[4];
 
     s3.deleteObject(
